refactor(plans): extrait FicheDuplicationService de DuplicatePlanService
Déplace la copie d'une fiche (création + budgets + annexes, notes via les
ficheFields) vers un FicheDuplicationService réutilisable, en vue de la
duplication d'une fiche seule. DuplicatePlanService.recreateFiches y
délègue et mappe l'erreur vers DUPLICATE_PLAN_ERROR. Aucun changement de
comportement : les e2e duplicate-plan restent verts sans modification.
