+++ apps/backend/src/plans/plans/duplicate-plan/duplicate-plan.service.ts
@@ -1,8 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { AddAnnexeService } from '@tet/backend/plans/fiches/add-annexe/add-annexe.service';
-import { CreateFicheService } from '@tet/backend/plans/fiches/create-fiche/create-fiche.service';
 import { FicheCreateAuthorization } from '@tet/backend/plans/fiches/create-fiche/fiche-create-authorization';
-import { FicheActionBudgetService } from '@tet/backend/plans/fiches/fiche-action-budget/fiche-action-budget.service';
+import { FicheDuplicationService } from '@tet/backend/plans/fiches/fiche-duplication/fiche-duplication.service';
 import ListFichesService from '@tet/backend/plans/fiches/list-fiches/list-fiches.service';
 import { PermissionService } from '@tet/backend/users/authorizations/permission.service';
 import { AuthenticatedUser } from '@tet/backend/users/models/auth.models';
@@ -18,12 +16,7 @@ import { UpsertPlanErrorEnum } from '../upsert-plan/upsert-plan.errors';
 import { UpsertPlanService } from '../upsert-plan/upsert-plan.service';
 import { DuplicatePlanError, DuplicatePlanErrorEnum } from './duplicate-plan.errors';
 import { DuplicatePlanInput } from './duplicate-plan.input';
-import {
-  AxeIdRemapping,
-  mapSourceFicheToDuplicate,
-} from './duplicated-fiche.mapper';
-import { mapSourceFicheAnnexes } from './duplicated-fiche-annexes.mapper';
-import { mapSourceFicheBudgets } from './duplicated-fiche-budgets.mapper';
+import { AxeIdRemapping } from './duplicated-fiche.mapper';
 
 const toPersonneId = (personne: Personne): PersonneId => ({
   tagId: personne.tagId,
@@ -64,9 +57,7 @@ export class DuplicatePlanService {
     private readonly listFichesService: ListFichesService,
     private readonly upsertPlanService: UpsertPlanService,
     private readonly upsertAxeService: UpsertAxeService,
-    private readonly createFicheService: CreateFicheService,
-    private readonly ficheBudgetService: FicheActionBudgetService,
-    private readonly addAnnexeService: AddAnnexeService
+    private readonly ficheDuplicationService: FicheDuplicationService
   ) {}
 
   async duplicate(
@@ -260,7 +251,7 @@ export class DuplicatePlanService {
     const ficheIdRemapping = new Map<number, number>();
 
     for (const source of actions) {
-      const created = await this.duplicateFiche({
+      const created = await this.ficheDuplicationService.duplicateFiche({
         source,
         collectiviteId,
         parentId: null,
@@ -268,7 +259,9 @@ export class DuplicatePlanService {
         axeIdRemapping,
         tx,
       });
-      if (!created.success) return created;
+      if (!created.success) {
+        return failure(DuplicatePlanErrorEnum.DUPLICATE_PLAN_ERROR);
+      }
       ficheIdRemapping.set(source.id, created.data);
     }
 
@@ -281,7 +274,7 @@ export class DuplicatePlanService {
         return failure(DuplicatePlanErrorEnum.DUPLICATE_PLAN_ERROR);
       }
 
-      const created = await this.duplicateFiche({
+      const created = await this.ficheDuplicationService.duplicateFiche({
         source,
         collectiviteId,
         parentId,
@@ -289,99 +282,12 @@ export class DuplicatePlanService {
         axeIdRemapping,
         tx,
       });
-      if (!created.success) return created;
+      if (!created.success) {
+        return failure(DuplicatePlanErrorEnum.DUPLICATE_PLAN_ERROR);
+      }
       ficheIdRemapping.set(source.id, created.data);
     }
 
-    return success(undefined);
-  }
-
-  private async duplicateFiche({
-    source,
-    collectiviteId,
-    parentId,
-    authorization,
-    axeIdRemapping,
-    tx,
-  }: {
-    source: FicheWithRelations;
-    collectiviteId: number;
-    parentId: number | null;
-    authorization: FicheCreateAuthorization;
-    axeIdRemapping: AxeIdRemapping;
-    tx: Transaction;
-  }): Promise<Result<number, DuplicatePlanError>> {
-    const { fiche, ficheFields } = mapSourceFicheToDuplicate({
-      source,
-      collectiviteId,
-      parentId,
-      axeIdRemapping,
-    });
-
-    const created = await this.createFicheService.createFicheWithAuthorization({
-      authorization,
-      fiche,
-      ficheFields,
-      tx,
-    });
-    if (!created.success) {
-      return failure(DuplicatePlanErrorEnum.DUPLICATE_PLAN_ERROR);
-    }
-    const newFicheId = created.data.id;
-
-    const budgetsResult = await this.copyBudgets(source, newFicheId, tx);
-    if (!budgetsResult.success) return budgetsResult;
-
-    const annexesResult = await this.copyAnnexes({
-      source,
-      newFicheId,
-      modifiedBy: authorization.user.id,
-      tx,
-    });
-    if (!annexesResult.success) return annexesResult;
-
-    return success(newFicheId);
-  }
-
-  private async copyAnnexes({
-    source,
-    newFicheId,
-    modifiedBy,
-    tx,
-  }: {
-    source: FicheWithRelations;
-    newFicheId: number;
-    modifiedBy: string;
-    tx: Transaction;
-  }): Promise<Result<undefined, DuplicatePlanError>> {
-    const sourceAnnexes =
-      await this.addAnnexeService.loadRawAnnexesForDuplication(
-        source.id,
-        source.collectiviteId,
-        tx
-      );
-    const annexes = mapSourceFicheAnnexes(sourceAnnexes, {
-      newFicheId,
-      collectiviteId: source.collectiviteId,
-      modifiedBy,
-    });
-    const result = await this.addAnnexeService.insertAnnexes(annexes, tx);
-    if (!result.success) {
-      return failure(DuplicatePlanErrorEnum.DUPLICATE_PLAN_ERROR);
-    }
-    return success(undefined);
-  }
-
-  private async copyBudgets(
-    source: FicheWithRelations,
-    newFicheId: number,
-    tx: Transaction
-  ): Promise<Result<undefined, DuplicatePlanError>> {
-    const budgets = mapSourceFicheBudgets(source, newFicheId);
-    const result = await this.ficheBudgetService.insertBudgets(budgets, tx);
-    if (!result.success) {
-      return failure(DuplicatePlanErrorEnum.DUPLICATE_PLAN_ERROR);
-    }
     return success(undefined);
   }
 }
